Refactored names to be Pascal
